backend/requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
jwt==1.4.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
rest-framework-simplejwt==0.0.2
sqlparse==0.5.4
tzdata==2025.2
uuid6==2025.0.1
